Put the "Reset User prefs" button in it's right place
